fixed bug with typization of decimal to float from sql connection.

+++ GameKeyCasino/GameCasino.BLL/UserLogic.cs
@@ -16,12 +16,12 @@ namespace GameCasino.BLL
             _userDao.Add(user);
         }
 
-        public void AddMoney(int idUser,float money)
+        public void AddMoney(int idUser,decimal money)
         {
             _userDao.AddMoney(idUser,money);
         }
 
-        public void RemoveMoney(int idUser,float money)
+        public void RemoveMoney(int idUser,decimal money)
         {
             _userDao.RemoveMoney(idUser,money);
         }
